Moved all handlers to individual files to comply with CodeFactor.
